fix: some database and dto field types

+++ domain/dtos/create-merchant-dto.go
@@ -2,5 +2,5 @@ package dtos
 
 type CreateMerchantDTO struct {
 	Name      string  `json:"name" binding:"required"`
-	Commision float32 `json:"commision" binding:"required,min=0,max=100"`
+	Commision float64 `json:"commision" binding:"required,min=0,max=100"`
 }
