+++ domain/dtos/update-merchant-dto.go
@@ -2,5 +2,5 @@ package dtos
 
 type UpdateMerchantDTO struct {
 	Name      string  `json:"name"`
-	Commision float32 `json:"commision" binding:"min=0,max=100"`
+	Commision float64 `json:"commision" binding:"min=0,max=100"`
 }
