+++ domain/entities/merchant.go
@@ -3,6 +3,6 @@ package entities
 type Merchant struct {
 	Model
 	Name         string        `json:"name" example:"walmart"`
-	Commision    float32       `json:"commision" example:"2.5"`
+	Commision    float64       `json:"commision" example:"2.5"`
 	Transactions []Transaction `json:"-"`
 }
