+++ domain/entities/transaction.go
@@ -3,7 +3,7 @@ package entities
 type Transaction struct {
 	Model
 	Ammount    uint64
-	Fee        float32
+	Fee        uint64
 	UserID     uint
 	MerchantID uint
 }
